requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=5.4.1=pypi_0
annotated-types=0.7.0=pypi_0
attrs=24.2.0=pypi_0
blinker=1.8.2=pypi_0
bzip2=1.0.8=h2466b09_7
ca-certificates=2024.8.30=h56e8100_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-ai-generativelanguage=0.6.10=pypi_0
google-api-core=2.21.0=pypi_0
google-api-python-client=2.149.0=pypi_0
google-auth=2.35.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.8.3=pypi_0
googleapis-common-protos=1.65.0=pypi_0
grpcio=1.66.2=pypi_0
grpcio-status=1.66.2=pypi_0
httplib2=0.22.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
libffi=3.4.2=h8ffe710_5
libsqlite=3.46.1=h2466b09_0
libzlib=1.3.1=h2466b09_2
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.1=pypi_0
mdurl=0.1.2=pypi_0
narwhals=1.9.3=pypi_0
numpy=2.1.2=pypi_0
openssl=3.3.2=h2466b09_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=pyh8b19718_1
proto-plus=1.24.0=pypi_0
protobuf=5.28.2=pypi_0
pyarrow=17.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
python=3.10.15=hfaddaf0_2_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
pytz=2024.2=pypi_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rich=13.9.2=pypi_0
rpds-py=0.20.0=pypi_0
rsa=4.9=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
streamlit=1.39.0=pypi_0
tenacity=9.0.0=pypi_0
tk=8.6.13=h5226925_1
toml=0.10.2=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
ucrt=10.0.22621.0=h57928b3_1
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
vc=14.3=h8a93ad2_22
vc14_runtime=14.40.33810=hcc2c482_22
vs2015_runtime=14.40.33810=h3bf8584_22
watchdog=5.0.3=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
xz=5.2.6=h8d14728_0
